chore: 클래스들 위치 조정
