+++ src/main/java/Auction_shop/auction/security/auth/service/AuthService.java
@@ -5,9 +5,9 @@ import Auction_shop.auction.domain.member.service.MemberService;
 import Auction_shop.auction.domain.refreshToken.RefreshToken;
 import Auction_shop.auction.domain.refreshToken.repository.RefreshTokenRepository;
 import Auction_shop.auction.security.jwt.JwtUtil;
-import Auction_shop.auction.web.dto.AuthRequestDto;
-import Auction_shop.auction.web.dto.AuthResponseDto;
-import Auction_shop.auction.web.dto.MemberResponseDto;
+import Auction_shop.auction.web.dto.auth.AuthRequestDto;
+import Auction_shop.auction.web.dto.auth.AuthResponseDto;
+import Auction_shop.auction.web.dto.member.MemberResponseDto;
 import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
