+++ src/main/java/Auction_shop/auction/web/dto/auth/AuthRequestDto.java
@@ -1,4 +1,4 @@
-package Auction_shop.auction.web.dto;
+package Auction_shop.auction.web.dto.auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,8 +9,6 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class AuthResponseDto {
-    private Long id;
-    private String accessToken;
-    private boolean available;
+public class AuthRequestDto {
+    String uuid;
 }
