+++ src/main/java/Auction_shop/auction/web/dto/member/MemberUpdateDto.java
@@ -1,4 +1,4 @@
-package Auction_shop.auction.web.dto;
+package Auction_shop.auction.web.dto.member;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
